Move safeGetline( file, line) out of class RbFileManager
The safeGetline( ) is not really connected to the RbFilemanager class.
It doesn't use any of the parts of the class.

However, because it was declared as a member of RbFilemanager,
you couldn't call safeGetline( ) without creating an empty RbFilemanager
object.

Move safeGetline( ) out of class RbFileManager to decrease dependencies
out that class, and to make the code cleaner.