Make sure we don't have any dangling pointers to the function frame.
If the function creates a variable that doesn't get built before the
function returns, build it then.

Later we could try refcounting the function frame, so that it is deleted
only when any vectors that need to get built using it are fully built.

But for now, lets do this.